refactor : folder structure -> with/without auth layouts
