+++ src/app/(with-auth)/app.tsx
@@ -4,7 +4,7 @@ import Script from "next/script";
 import dynamic from "next/dynamic";
 import { useUser } from "@clerk/nextjs";
 import { Analytics } from "@vercel/analytics/react";
-import { useCustomScrollbar } from "./hooks/useCustomScrollbar";
+import { useCustomScrollbar } from "../hooks/useCustomScrollbar";
 import { useMemo, useState, useContext, createContext } from "react";
 import type {
   NotificationPollLikeItem,
@@ -12,7 +12,7 @@ import type {
   NotificationCommentLikeItem,
   NotificationFollowPendingItem,
   NotificationFollowAcceptedItem,
-} from "./components/Header/actions";
+} from "../components/Header/actions";
 
 const NotificationsHandler = dynamic(
   () =>
