+++ src/app/(with-auth)/polls/[id]/page.tsx
