+++ src/app/(with-auth)/polls/create/page.tsx
