+++ src/app/(with-auth)/users/@votes/[id]/page.tsx
