+++ src/app/(with-auth)/users/@votes/default.tsx
