+++ src/app/(with-auth)/users/components/FollowButtonClient.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { Loader } from "@/app/components/Loader";
-import { follow, unfollow } from "@/app/users/actions";
+import { follow, unfollow } from "@/app/(with-auth)/users/actions";
 
 export function FollowButtonClient({
   userId,
@@ -27,7 +27,7 @@ export function FollowButtonClient({
 
   return (
     <button
-      className="hovact:bg-neutral-800 h-fit w-fit rounded-lg bg-neutral-900 px-1.5 py-0.5 transition-colors"
+      className="h-fit w-fit rounded-lg bg-neutral-900 px-1.5 py-0.5 transition-colors hovact:bg-neutral-800"
       onClick={handleClick}
     >
       {isFollowing ? "unfollow" : "follow"}
