+++ src/app/(with-auth)/users/components/user-followed/ActiveFollowedCard.tsx
