+++ src/app/(with-auth)/welcome/Confetti.tsx
