--- src/app/privacy-policy/page.tsx
@@ -1,113 +1,0 @@
-export default function PrivacyPolicyPage() {
-  return (
-    <main className="gap-2 [&>h1]:text-5xl [&>h1]:font-bold [&>h2]:pt-10 [&>h2]:text-3xl [&>h2]:font-semibold [&>p]:py-2 [&>ul]:list-disc [&>ul]:pl-6">
-      <h1>Privacy Policy</h1>
-      <p className="text-neutral-200">Last updated: February 19, 2024</p>
-      <p>
-        This Privacy Policy describes how Pollist (&quot;we,&quot;
-        &quot;us,&quot; or &quot;our&quot;) collects, uses, and shares
-        information about you through our mobile application (the
-        &quot;App&quot;). By using the App, you agree to the collection, use,
-        and sharing of your information as described in this Privacy Policy.
-      </p>
-      <h2>Information We Collect</h2>
-      <p>
-        User Provided Information: We collect information you provide directly
-        to us when you create an account, such as your username, email address,
-        and other contact or profile information you choose to provide.
-      </p>
-      <p>
-        Automatically Collected Information: When you use our App, we
-        automatically collect certain information, including your IP address,
-        device type, operating system, and usage details (such as the time and
-        frequency of your use of the App).
-      </p>
-      <h2>How We Use Your Information</h2>
-      <p>We use the information we collect to:</p>
-      <ul>
-        <li>Provide, maintain, and improve our App;</li>
-        <li>
-          Communicate with you about products, services, and promotional offers;
-        </li>
-        <li>
-          Monitor and analyze trends, usage, and activities in connection with
-          our App;
-        </li>
-        <li>
-          Detect, investigate, and prevent fraudulent transactions and other
-          illegal activities and protect the rights and property of Pollist and
-          others.
-        </li>
-      </ul>
-      <h2>Sharing of Information</h2>
-      <p>We may share your information as follows:</p>
-      <ul>
-        <li>
-          With vendors, consultants, and other service providers who need access
-          to such information to carry out work on our behalf;
-        </li>
-        <li>
-          In response to a request for information if we believe disclosure is
-          in accordance with, or required by, any applicable law, regulation, or
-          legal process;
-        </li>
-        <li>
-          Between and among Pollist and our current and future parents,
-          affiliates, subsidiaries, and other companies under common control and
-          ownership;
-        </li>
-        <li>With your consent or at your direction.</li>
-      </ul>
-      <p>
-        Additionally, we may share aggregated or de-identified information,
-        which cannot reasonably be used to identify you.
-      </p>
-      <h2>Advertising and Analytics Services</h2>
-      <p>
-        We may allow others to provide analytics services and serve
-        advertisements on our behalf across the web and in mobile applications.
-        These entities may use cookies, web beacons, device identifiers, and
-        other technologies to collect information about your use of the App and
-        other websites and applications, including your IP address, web browser,
-        mobile network information, pages viewed, time spent on pages or in
-        apps, links clicked, and conversion information. This information may be
-        used by Pollist and others to, among other things, analyze and track
-        data, determine the popularity of certain content, deliver advertising
-        and content targeted to your interests on our services and other
-        websites, and better understand your online activity.
-      </p>
-      <h2>Your Choices</h2>
-      <p>
-        Account Information: You may update, correct, or delete information
-        about you at any time by logging into your account and modifying your
-        information or by contacting us.
-      </p>
-      <p>
-        Cookies: Most web browsers are set to accept cookies by default. If you
-        prefer, you can usually choose to set your browser to remove or reject
-        browser cookies.
-      </p>
-      <p>
-        Promotional Communications: You can opt out of receiving promotional
-        emails from us by following the instructions in those emails. If you opt
-        out, we may still send you non-promotional emails, such as those about
-        your account or our ongoing business relations.
-      </p>
-      <h2>Data Security</h2>
-      <p>
-        We take reasonable measures to help protect information about you from
-        loss, theft, misuse, and unauthorized access, disclosure, alteration,
-        and destruction.
-      </p>
-      <h2>Changes to the Privacy Policy</h2>
-      <p>
-        We may update this Privacy Policy from time to time. If we make changes,
-        we will notify you by revising the date at the top of the policy and, in
-        some cases, we may provide you with additional notice (such as adding a
-        statement to our homepage or sending you a notification).
-      </p>
-      {/* <h2>8. Contact Us</h2>
-        <p>If you have any questions about this Privacy Policy, please contact us at: [Insert Contact Information].</p> */}
-    </main>
-  );
-}
